requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.2
djangorestframework==3.11.0
httpie==2.0.0
idna==2.8
packaging==20.1
Pillow==7.0.0
pip-review==1.0
Pygments==2.5.2
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
